requirements.txt
python-dotenv
flask==0.12.2
gunicorn==19.7.1
requests==2.18.4